Persist the private pEPA the strategy engine already computes
Scouted 36 match entries and 12 pit entries for a real event, opened Strategy,
and watched it blend a correct org-private pEPA — then throw it away. Every
snapshot insert failed with

  invalid input syntax for type integer: "3.4"

because `private_epa_snapshots.scout_sample` is an integer column while the model
produces a quality-weighted sample (3.4 observations, not 3). The `catch {}` around
the loop hid it, so /api/org/analytics/private-epa, the pick desk's pEPA column and
the `strategy.private_edge` agent tool answered "No org pEPA snapshots yet" no
matter how much a team scouted — the differentiator docs/FEATURE_MAP.md leads with.

0518 widens the column to numeric(8,2) so the stored sample carries the weighting
the model used; both readers now cast it to float like their neighbours.

The failure was invisible for a second reason worth fixing structurally: withRls
puts the whole request in one transaction, so the swallowed error also aborted the
pit-signal and live-alert writes that follow with "current transaction is aborted".
Adds `withSavepoint` to @vantage/db and uses it for both best-effort blocks, so an
optional write that fails costs only itself.

Verified end to end on the local database: 36 scout entries -> Strategy ->
private_epa_snapshots row (scout_sample 3.40) -> /api/org/analytics/private-epa
returns it -> the pick desk shows pEPA 61.02 for the team with n>=3 and null for
the teams with fewer, which is the honest behaviour.

+++ packages/agent/src/tools.ts
@@ -551,7 +551,7 @@ export function createVantageToolRegistry(): AIToolRegistry {
           try {
             const snapshots = await client.query(
               `SELECT team_key AS "teamKey", public_epa::float AS "publicEpa", pepa::float AS pepa,
-                      scout_component_epa::float AS "scoutComponentEpa", scout_sample AS "scoutSample",
+                      scout_component_epa::float AS "scoutComponentEpa", scout_sample::float AS "scoutSample",
                       components, computed_at::text AS "computedAt"
                FROM private_epa_snapshots
                WHERE org_id = $1 AND event_key = $2
